Prototype reactive block communication
Only a few blocks are currently functional - most tests are failing for
now.

BlockInput is now a thing of the past.  Using the "most" reactive
programming library we now declare block connections up front via
reactive streams.  Drawables mark their streams so that they trigger
redraw.  This means that all processing is done immediately after a
change occurs, and that change organically only affects the blocks
subscribed to that stream.  When and if a drawable block's inputs
change, the visualization is redrawn by simply enumerating objects to be
drawn and passing them to THREE.js's WebGLRenderer.